+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   * Newer version available window not appearing in Flathub version.
   * On very fast Therion executions, therion.log file content not being included in Therion run output.
   * Bézier created from straight lines are too tight.
+  * Status bar info for selected elements not internationalized.
 * Infrastructure maintenance:
   * Enabling github created flatpak generation.
   * New "New version available" page.